dialog on register successful
