+++ app/src/main/java/com/example/erabook/firebaseActivities/LogIn.kt
@@ -1,62 +1,58 @@
-package com.example.erabook.firebaseServices
+package com.example.erabook.firebaseActivities
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
-import com.example.erabook.databinding.ActivityRegisterBinding
+import com.example.erabook.databinding.ActivityLogInBinding
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 
-class Register : AppCompatActivity() {
-    private lateinit var binding: ActivityRegisterBinding
+class LogIn : AppCompatActivity() {
+    private lateinit var binding: ActivityLogInBinding
     private lateinit var auth: FirebaseAuth
 
-//    public override fun onStart() {
+    //    public override fun onStart() {
 //        super.onStart()
 //        val currentUser = auth.currentUser
 //        if (currentUser != null) {
 //            TODO("open the fragment that shows the user profile")
 //        }
 //    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityRegisterBinding.inflate(layoutInflater)
+        binding = ActivityLogInBinding.inflate(layoutInflater)
         setContentView(binding.root)
+
         auth = Firebase.auth
 
 
-
         binding.apply {
-            register.setOnClickListener {
-                val email = emailInput.editText?.text.toString()
-                val password = passwordInput.editText?.text.toString()
-
-                if (email.isEmpty()) {
+            loginButton.setOnClickListener {
+                val emailLogIn = emailLoginInput.editText?.text.toString()
+                val passwordLogIn = passwordLoginInput.editText?.text.toString()
+                if (emailLogIn.isEmpty()) {
                     Toast.makeText(
                         baseContext,
                         "Enter email please!",
                         Toast.LENGTH_SHORT
                     ).show()
-                    return@setOnClickListener
                 }
-                if (password.isEmpty()) {
+                if (passwordLogIn.isEmpty()) {
                     Toast.makeText(
                         baseContext,
                         "Enter password please!",
                         Toast.LENGTH_SHORT
                     ).show()
-                    return@setOnClickListener
                 }
-                auth.createUserWithEmailAndPassword(email, password)
+                auth.signInWithEmailAndPassword(emailLogIn.trim(), passwordLogIn)
                     .addOnCompleteListener { task ->
                         if (task.isSuccessful) {
                             Toast.makeText(
                                 baseContext,
-                                "Account created.",
-                                Toast.LENGTH_SHORT
+                                "Log in successful.",
+                                Toast.LENGTH_SHORT,
                             ).show()
                         } else {
                             Toast.makeText(
@@ -68,11 +64,11 @@ class Register : AppCompatActivity() {
                     }
 
             }
-            alreadyHaveLogin.setOnClickListener {
-                val loginIntent = Intent(baseContext, LogIn::class.java)
-                startActivity(loginIntent)
+            goToRegister.setOnClickListener {
+                val registerIntent = Intent(baseContext, Register::class.java)
+                startActivity(registerIntent)
             }
-            backRegister.setOnClickListener {
+            backLogin.setOnClickListener {
                 finish()
                 super.onBackPressedDispatcher
             }
