+++ app/src/main/res/layout/item_book.xml
@@ -8,7 +8,7 @@
     android:layout_margin="5dp"
     >
     <androidx.cardview.widget.CardView
-        android:id="@+id/signup_successful"
+        android:id="@+id/favorite_item"
         android:layout_width="150dp"
         android:layout_height="250dp"
         app:layout_constraintTop_toTopOf="parent"
