+++ app/src/main/res/layout/item_favorite_book.xml
@@ -10,7 +10,7 @@
     </data>
 
     <com.google.android.material.card.MaterialCardView
-        android:id="@+id/signup_successful"
+        android:id="@+id/favorite_item"
         android:layout_width="150dp"
         android:layout_height="250dp"
         app:cardCornerRadius="14dp"
