+++ app/src/main/res/layout/item_genre.xml
@@ -4,7 +4,7 @@
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <androidx.cardview.widget.CardView
-        android:id="@+id/signup_successful"
+        android:id="@+id/favorite_item"
         android:layout_width="150dp"
         android:layout_height="120dp"
         app:layout_constraintTop_toTopOf="parent"
