--- app/src/main/res/layout/signup_successful.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <androidx.cardview.widget.CardView
-        android:id="@+id/signup_successful"
-        android:layout_width="match_parent"
-        android:layout_height="400dp"
-        android:layout_margin="40dp"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:cardCornerRadius="24dp"
-        app:cardElevation="2dp"
-        app:cardUseCompatPadding="true"
-        >
-        <TextView
-            android:id="@+id/signupsucc"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:padding="@dimen/space20"
-            android:textStyle="bold"
-            android:text="@string/signup_succ"
-            android:textSize="34sp"
-            android:textColor="@color/primary"
-            android:layout_margin="@dimen/space20"
-            android:layout_gravity="top"
-            />
-
-        <TextView
-            android:id="@+id/signupsuccdesc"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
-            android:layout_margin="@dimen/space20"
-            android:padding="@dimen/space20"
-            android:text="@string/signup_succ_desc"
-            android:textSize="14sp"
-            android:textStyle="bold" />
-
-        <ImageView
-            android:layout_width="100dp"
-            android:layout_height="100dp"
-            android:layout_gravity="bottom|center"
-            android:src="@drawable/successful"
-            android:layout_margin="40dp"
-            android:contentDescription="@string/desc_img"
-            />
-    </androidx.cardview.widget.CardView>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
